automationproject/requirements.txt
praw
langchain
langchain_community
langgraph
openai
langchain_openai
langchain_groq
IPython
pandas
tweepy
supabase
fastapi
anthropic
openpyxl
python-pptx
seaborn
matplotlib
IPython
